\name{GAIPE-package}
\alias{GAIPE-package}
\alias{GAIPE}
\docType{package}
\title{
Graphical Extension with Accuracy in Parameter Estimation (GAIPE)
}
\description{
GAIPE implements graphical representation of accuracy in parameter estimation (AIPE) on RMSEA for sample size planning in structural equation modeling. Sample sizes suggested by RMSEA with AIPE method and power analysis approach can also be obtained separately using the provided functions.
}
\details{
\tabular{ll}{
Package: \tab GAIPE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-08\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Tzu-Yao Lin
Maintainer: Tzu-Yao Lin <zaiyaolin@gmail.com>
}
\references{
Lin, T.-Y. & Weng, L.-J. (Under review). Graphical extension of sample size planning with AIPE on RMSEA. \emph{Structural Equation Modeling: A Multidisciplinary Journal.}
}
