\name{GAIPE-package}
\alias{GAIPE-package}
\alias{GAIPE}
\docType{package}
\title{
Graphical Extension with Accuracy in Parameter Estimation (GAIPE)
}
\description{
Implements graphical extension with accuracy in parameter estimation (AIPE) on RMSEA for sample size planning in structural equation modeling based on Lin, T.-Z. & Weng, L.-J. (2014) <doi: 10.1080/10705511.2014.915380>.
}
\details{
\tabular{ll}{
Package: \tab GAIPE\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2022-05-24\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Tzu-Yao Lin
Maintainer: Yao Lin <zaiyaolin@gmail.com>
}
\references{
Lin, T.-Z. & Weng, L.-J. (2014) Graphical Extension of Sample Size Planning With AIPE on RMSEA Using R. Structural Equation Modeling, 21, 482-490. doi: 10.1080/10705511.2014.915380
}
